Move reusable part in server.js
